prepare for release

+++ core/src/in/srain/cube/cache/IFileCache.java
@@ -8,7 +8,7 @@ public interface IFileCache {
 
     public void clearCache();
 
-    public int getMaxSize();
+    public long getMaxSize();
 
     public boolean has(String key);
 }
